% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A549_MAX_gr.R
\docType{data}
\name{A549_MAX_gr}
\alias{A549_MAX_gr}
\title{A549_MAX_gr}
\format{
\subsection{\code{A549_MAX_gr}}{

The GRange file of A549 MAX ChIP-Seq data
\describe{
\item{cell line}{A549}
\item{year}{2012}
}
}
}
\source{
\url{https://www.http://genome.ucsc.edu/cgi-bin/hgFileUi?db=hg19&g=wgEncodeAwgTfbsUniform}
}
\usage{
A549_MAX_gr
}
\description{
An example of the A549_MAX_gr file for function 'bed_input'
}
\keyword{datasets}
