% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_input.R
\name{fasta_input}
\alias{fasta_input}
\title{Read and process DNA sequence from a FASTA file}
\usage{
fasta_input(fasta_file, shapeind = "all")
}
\arguments{
\item{fasta_file}{path to the FASTA file containing the DNA sequence.}

\item{shapeind}{integer, the index of shape feature to process, default is "all". 1 to 5 indicate feature "MGW", "HelT", "ProT", "Roll", and "EP" respectively.}
}
\value{
An array of processed shape data.
}
\description{
This function reads a DNA sequence from a FASTA file and processes it to the DNA shape data. The DNA shape data can be used as the input of function SMEM() or SMEM_Gibbs() for DNA shape motif discovery
}
