% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_motif_merge.R
\name{fasta_motif_merge}
\alias{fasta_motif_merge}
\title{Count and merge DNA shape motif location given by Expectation–Maximization algorithm}
\usage{
fasta_motif_merge(
  motif_location_array,
  fasta_file,
  motifLength = NULL,
  filename = "location_merge"
)
}
\arguments{
\item{motif_location_array}{the motif location array returned by function 'SMEM'.}

\item{fasta_file}{the fasta file used for motif discovery.}

\item{motifLength}{integer, length of shape motif to be discovered, default is 12.}

\item{filename}{character, name of output file, default is "location_merge".}
}
\value{
A list of motif location data frame and motif shape data
}
\description{
This function reads the location array of shape motif and output final motif location results. The motif discovery with BED file input can get the genome location of the motif. The FASTA input will get results only based on the FASTA file.
}
