% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_input.R
\name{gr_input}
\alias{gr_input}
\title{Read and process DNA sequence from a GRanges file}
\usage{
gr_input(
  gr_file,
  peakLength = 100,
  total_peak_number = 100,
  fasta_name = "fasta.fa",
  shapeind = "all",
  sort_by = "signalValue",
  sort_descend = TRUE
)
}
\arguments{
\item{gr_file}{The GRanges file containing the peak information.}

\item{peakLength}{integer, length of center peak to be processed, default is 100.}

\item{total_peak_number}{integer, total peak number, default is 100.}

\item{fasta_name}{file name of the output FASTA file, default is "fasta.fa".}

\item{shapeind}{integer, the index of shape feature to process, default is "all". 1 to 14 indicate feature '1.MGW', '2.HelT', '3.ProT', '4.Roll', '5.EP', '6.Rise', '7.Shift', '8.Slide', '9.Tilt', '10.Buckle', '11.Opening', '12.Shear', '13.Stagger', '14.Stretch' respectively.}

\item{sort_by}{parameter used to sort the gr file, default is "signalValue".}

\item{sort_descend}{sort the peaks in a descending order according to 'sort_by', default is "TRUE".}
}
\value{
An array of processed shape data
}
\description{
This function reads a DNA sequence from a GRanges file and processes it to the DNA shape data. The DNA shape data can be used as the input of function SMEM() or SMEM_Gibbs() for DNA shape motif discovery.
}
