% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_features_visualizer.R, R/motif_visualizer.R
\name{motif_visualizer}
\alias{motif_visualizer}
\title{Visualize the DNA shape motif output by the merge functions}
\usage{
motif_visualizer(
  motif_data_frame,
  motif_shape_data,
  motifFrequency = 0.6,
  shapeIndex = NULL,
  motifLength = NULL,
  align_method = "pearson",
  ref_number = 1,
  input_form = NULL,
  continue_alignment = NULL,
  top_motif = NULL,
  threshold = NULL
)

motif_visualizer(
  motif_data_frame,
  motif_shape_data,
  motifFrequency = 0.6,
  shapeIndex = NULL,
  motifLength = NULL,
  align_method = "pearson",
  ref_number = 1,
  input_form = NULL,
  continue_alignment = NULL,
  top_motif = NULL,
  threshold = NULL
)
}
\arguments{
\item{motif_data_frame}{the motif motif_data_frame returned by function 'fasta/bed_motif_merge'.}

\item{motif_shape_data}{the motif motif_shape_data returned by function 'fasta/bed_motif_merge'.}

\item{motifFrequency}{integer, the threshold for selected motifs with top exsitence frequency.}

\item{shapeIndex}{the index of shape feature to process, 1 to 5 indicate feature "MGW", "HelT", "ProT", "Roll", and "EP" respectively. "all" for mutiple shape alignment.}

\item{motifLength}{integer, length of shape motif to be aligned and visualized.}

\item{align_method}{character, the metric used for alignment, default is "pearson", can also choose "distance".}

\item{ref_number}{integer, the reference motif for alignment, default is 1.}

\item{input_form}{integer, '0' for 'fasta_motif_merge' or '1' for 'bed_motif_merge', user input while the function is running.}

\item{continue_alignment}{integer, '0' for 'No' or '1' for 'Yes', user input while the function is running.}

\item{top_motif}{integer, number of motifs to align, user input while the function is running.}

\item{threshold}{double, the threshold for align method 'pearson', user input while the function is running.}
}
\value{
A list of motif location data frame and motif shape data

A list of motif location data frame and motif shape data
}
\description{
This function reads the motif shape data and the motif data frame

This function reads the motif shape data and the motif data frame
}
\details{
This function output possible shape motifs shape feature and sequence logo

This function output possible shape motifs shape feature and sequence logo
}
